refactor: legacy ingestion cleanup, max_tokens config, test reorganisation
Closes #1532 — remove 3 legacy ingestion files replaced by unified
pipeline (src/ingestion/{docling_client,gdrive_flow,service}.py,
~1 313 lines). Clean up corresponding per-file-ignores in pyproject.toml.

Closes #1537 — make LLMService max_tokens config-driven: parameter type
changed from int to int | None; fallback to _DEFAULT_MAX_TOKENS = 4096
(matches GraphConfig.llm_max_tokens). Callers that pass a config value
now fully control the limit; default behaviour is unchanged.

Closes #1552 — add TEST_SERVICES_HOST fixture + env var to
tests/integration/conftest.py so integration tests can target a remote
Docker daemon (e.g. MacBook). Move 8 mock-only integration tests to
tests/unit/ where they belong:
  test_llm_generate, test_contextual_integration,
  test_contextualized_pipeline, test_colbert_backfill, test_redis_url,
  test_qdrant_service, test_service_chain_agent_tools,
  test_voyage_integration
